requirements.txt
torch
pyyaml
tensorboard
nuscenes-devkit
numpy
dill
matplotlib
sqlitedict